Add support for Sonata PCB
This commit assumes revision 0.2 of the board. It adds a pin file, a top
level module, a clock generator and the appropriate changes to the core
file. It also updates the getting started guide to use the Sonata PCB by
default.

This commit is derived from a PR in the demo system:
https://github.com/lowRISC/ibex-demo-system/pull/86