Generate GPIO, UART, I2C, & SPI based on params
This creates a sonata_xbar_main module wrapper around xbar_main that can
eventually be auto-generated based on the number of GPIO, SPI, I2C and UART
blocks. It also puts the block numbers in a package that can eventually
be generated. The commit keeps the interrupt indees unchanged to avoid
software toil.