Change boot_rom and regenerate crossbars
Boot ROM is not at address zero, so the more involved code requiring
a second scratch register is required.
Boot ROM code attempts to store to offsets within the first half
of the debug device page indicate the status to the debug module,
so both data and instruction ports must be able to access the
device.